Fix comparison
